src/serving/inference_service/requirements.txt
fastapi
uvicorn
numpy
pillow
prometheus-fastapi-instrumentator
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
torchvision==0.20.1
prometheus-client
alibi-detect
transformers